REFACTOR: separated api layer from app layer, aupgraded dynamic filters for logs and configs, added dynamic front-end fields resolution, rewrited filtering in logs, renamed interfaces according to best go practices

+++ watcher/watcher.go
@@ -29,7 +29,7 @@ type UploadWatcherParams struct {
 	Period       int
 	NextUploadOn time.Time
 	LastLogId    int
-	DomainId     int
+	DomainId     int64
 	UserId       int
 }
